Rewriting the diff algo
This should allow us to provide better output, and also be smart
about extra wrapping elements.

Adding in weights (ie. "how different") and returning them from 
the diff means we can attempt the actual child as the expected,
and see if it is "better" (ie. has a lower weight).

This might work both ways around, so if there should be a wrapper
but there isn't one, or if an extra wrapper has been added. This may be 
too expensive and/or complicated - initial aim is just ignoring extra 
wrappers in "actual".  This fits the primary use-case of HoC for react.