Fix `queried for` to return correct element when used with promises
`'queried for', ...).then(...` was not returning the public instance of the
component or element.

Tests amended - note that the tests were wrong also, as they tested the result
using `'to have rendered'`, which would not be possible with an HTML element
(at least not until we have an unexpected-htmllike adapter for the DOM).

Closes #26