Fix class comparison for `on` clauses in events
Make all the options the same across alll assertions based on the flags.
Previously there was a mismatch between (for example) `queried for` flags and `to have rendered` flags. This is now extracted, so they are always calculated the same.

A nice extension to this would be to be able to pass the htmllike options directly in the assertion, if a special combination was required.