Add support for `react-test-renderer`
The same assertions are available for checking rendering,
querying, calling events as for the shallow renderer.

In order to use it, require `unexpected-react/test-renderer`. The test-renderer cannot be used in conjunction with the DOM or
shallow renderer, so it is not possible to require 
`unexpected-react` in the same test run. For mocha, you need
to have separate runs if you have both sorts of tests.