Change the `multiple` test to check the JSON output
The tests are run in parallel, so they can appear in any order. 
For the `multiple` test, that has more than one test file, we
use the JSON output of jest, and run it through a normalizer
to remove or "normalize" anything that is different between 
runs (e.g. `startTime`, absolute paths)