Fix comments in snapshots
Snapshots were being written with comments with raw JSON rather than
the JSX. They were being identified as the RawReactTestRendererJson
when `inspect` ran, which meant the raw JSON was outputted. Now they
are identified correctly.

I think because of the new hierarchy, it now actually outputs
correctly in the test, but we need @sunesimonsen's fix to properly
fix this - I'm not sure what combination is actually making this work
- the RawReactTestRendererJson doesn't have an `inspect` method.